+++ packages/babel-plugin/src/types.ts
@@ -107,6 +107,9 @@ export interface PluginOptions {
   /**
    * Adds a defined prefix to the generated classes' hashes.
    * Useful in micro frontend environments to avoid clashing/specificity issues.
+   *
+   * Avoid mixing this with extraction as this may throw an error if combined with extraction
+   * or `extract: true` in Webpack loaders or Parcel tranformers.
    */
   classHashPrefix?: string;
 }